add auth

+++ app/Http/Controllers/TasksController.php
@@ -15,13 +15,25 @@ class TasksController extends Controller
      */
     public function index()
     {
-         // メッセージ一覧を取得
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = Task::orderBy('created_at', 'desc')->paginate(10);
 
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index', $data);
+        }else{
+         // Welcomeビューでそれらを表示
+        return view('welcome');
+        }
+        
+        
     }
 
     /**
@@ -51,13 +63,14 @@ class TasksController extends Controller
             'status' => 'required|max:10',   // 追加
             'content' => 'required|max:255',
         ]);
-                // メッセージを作成
-        $task = new Task;
-        $task->status = $request->status;    
-        $task->content = $request->content;
-        $task->save();
+              
+         // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' =>  $request->status,
+            'content' => $request->content,
+        ]);
 
-        // トップページへリダイレクトさせる
+         // トップページへリダイレクトさせる
         return redirect('/');
     }
 
@@ -130,10 +143,12 @@ class TasksController extends Controller
     {
          // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        
+       // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
-        // トップページへリダイレクトさせる
         return redirect('/');
     }
 }
